refactor(ui): move all routes into [lng] folder to enable i18n for routes
