+++ app/src/app/[lng]/(auth)/check-email/page.tsx
