+++ app/src/app/[lng]/(auth)/forgot-password/page.tsx
