+++ app/src/app/[lng]/(auth)/layout.tsx
