+++ app/src/app/[lng]/(auth)/reset-successful/page.tsx
