+++ app/src/app/[lng]/(auth)/signup/page.tsx
