+++ app/src/app/[lng]/layout.tsx
@@ -1,0 +1,33 @@
+import '../globals.css'
+import type { Metadata } from 'next'
+import { Providers } from '../providers'
+import { dir } from 'i18next'
+import { languages } from '@/i18n/settings'
+
+export const metadata: Metadata = {
+  title: 'CityCatalyst',
+  description: 'Make building a climate inventory a breeze',
+}
+
+export async function generateStaticParams() {
+  return languages.map((lng: string) => ({ lng }));
+}
+
+export default function RootLayout({
+  children,
+  params: { lng }
+}: {
+  children: React.ReactNode,
+  params: { lng: string }
+}) {
+  return (
+    <html lang={lng} dir={dir(lng)}>
+      <body>
+        <Providers>
+          {children}
+        </Providers>
+      </body>
+    </html>
+  )
+}
+
