+++ app/src/app/[lng]/onboarding/page.tsx
