+++ app/src/app/[lng]/onboarding/setup/page.tsx
