+++ app/src/app/[lng]/page.tsx
@@ -1,21 +1,21 @@
 'use client'
 
 import { NavigationBar } from '@/components/navigation-bar'
-import { Button, Heading, Text } from '@chakra-ui/react'
+import { Button, Heading } from '@chakra-ui/react'
 import Link from 'next/link'
 
-export default function Home() {
+export default function Home({ params: { lng } }: { params: { lng: String } }) {
   return (
     <main>
       <NavigationBar />
       <div className="p-24 flex flex-col space-y-4 items-start">
         <Heading>Dashboard</Heading>
-        <Link href="/login" passHref legacyBehavior>
+        <Link href={`/${lng}/login`} passHref legacyBehavior>
           <Button as="a" w={32}>
             Login
           </Button>
         </Link>
-        <Link href="/signup" passHref legacyBehavior>
+        <Link href={`/${lng}/signup`} passHref legacyBehavior>
           <Button as="a" w={32}>
             Sign Up
           </Button>
