--- app/src/app/layout.tsx
@@ -1,25 +1,0 @@
-import './globals.css'
-import type { Metadata } from 'next'
-import { Providers } from './providers'
-
-export const metadata: Metadata = {
-  title: 'CityCatalyst',
-  description: 'Make building a climate inventory a breeze',
-}
-
-export default function RootLayout({
-  children,
-}: {
-  children: React.ReactNode
-}) {
-  return (
-    <html lang="en">
-      <body>
-        <Providers>
-          {children}
-        </Providers>
-      </body>
-    </html>
-  )
-}
-
